Move core plugin types to synthesis utils
